showing the user his selected subjects
