+++ src/main/resources/templates/subjectSelection.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="https://www.thymeleaf.org">
 <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+    <meta charset="UTF-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css"
           integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
@@ -23,11 +23,10 @@
     <script src="https://code.jquery.com/jquery-1.10.2.js"></script>
     <title>Subject Selection</title>
 </head>
-<!--Subjects from user anzeigen, änhlich wie bei usertyps und datenbank-->
 <body>
 <div id="nav-placeholder"></div>
 <div class="container d-flex flex-column flex-md-col justify-content-between col-sm-12 col-md-6 col-lg-4">
-    <form id="subjectForm" th:action="@{/subjectSelection/save}" th:object="${subjects}" method="post">
+    <form id="subjectForm" th:action="@{/userSubject/save}" th:object="${subjects}" method="post">
         <div class="mx-auto" style="height: auto">
             <div class="form-group">
                 <h3>Select your Subject and School Year:</h3>
@@ -53,3 +52,4 @@
 <div id="footer-placeholder"></div>
 </body>
 </html>
+
