+++ src/main/resources/templates/userSubjects.html
@@ -11,7 +11,7 @@
 <div class="container">
     <h1>Your Selected Subjects</h1>
     <p>Year: <span th:text="${year}"></span></p>
-    <p>User ID: <span th:text="${userId}"></span></p>
+    <p>User ID: <span th:text="${user}"></span></p>
     <table class="table table-bordered">
         <thead>
         <tr>
